test: split incremental test cases to run in parallel
